Started working on edge buffer initialization

+++ src/ContentLoading.h
@@ -10,7 +10,7 @@
 
 #include "Renderable.h"
 
-struct PackedVertex{
+struct PackedVertex {
 	glm::vec3 position;
 	glm::vec2 uv;
 	glm::vec3 normal;
@@ -42,7 +42,7 @@ private:
 		std::vector<glm::vec3> & out_vertices, 
 		std::vector<glm::vec2> & out_uvs,
 		std::vector<glm::vec3> & out_normals,
-		std::vector<GLuint> &out_faces,
+		std::vector<GLushort> &out_faces,
 		std::vector<glm::vec3> & raw_verts);
 };
